requirements.txt
# Audio Processing
librosa>=0.9.0
soundfile>=0.11.0
pydub>=0.25.1

# Deep Learning
tensorflow>=2.10.0
keras>=2.10.0

# Machine Learning
scikit-learn>=1.1.0
scipy>=1.9.0

# Data Processing
numpy>=1.23.0
pandas>=1.5.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.10.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.64.0
joblib>=1.2.0

# Jupyter
jupyter>=1.0.0
notebook>=6.5.0
ipywidgets>=8.0.0

# Optional: Audio format conversion
# ffmpeg-python>=0.2.0
